requirements.txt
playwright==1.43.0
supabase==2.3.3
python-dotenv==1.0.1